refactor: make a `Walk` trait
